Updating PIN  tool to use long variables, instead of ints.
This allows the tool to be built on 64-bit platforms and where
strict warnings are enabled, since otherwise, ints would be 32
bits and couldn't store the full value of various pointers.

+++ tools/PINtool/src/region.cpp
@@ -4,8 +4,8 @@ bool region_monitoring_enabled = false;
 
 bool region_contains(region_t *region, void* address)
 {
-	char* region_start = (int)(region->start) + (char*)(region->loaded_address);
-    char* region_end = (int)(region->end) + (char*)(region->loaded_address);
+	char* region_start = (long)(region->start) + (char*)(region->loaded_address);
+    char* region_end = (long)(region->end) + (char*)(region->loaded_address);
 
     if(region_start <= address && address < region_end)
     {
